add USER command and limit for it

+++ Server/modules/command/QUIT.js
@@ -1,4 +1,10 @@
 
 module.exports = function(socket, command) {
+
+    if(!socket.client.identity) {
+        socket.send(':'+config.ip+' 451 * QUIT :You have not registered');
+        return;
+    }
+
     socket.client.delete();
 }
